processor: fix upload of large files to zenodo
The Zenodo API has a problem when uploading large files with POST, so we now use
PUT for the upload. I reported this issue to Zenodo as well.

Using PUT is a little ugly since unirest does not support PUT with file as a
body, so we have to directly use the apache http client.